moved data and random utils to utils folder
